info-leaks/reading: Fix broken CNS repo links in README.md
The links referring to the CNS repository in the README.md file are
broken. This causes a 404 - page not found error. This commit fixes the
issue by replacing the broken links with the links to the precompiled
binaries in the CNS GitHub repository. Unnecessary link titles related to
the CNS repository were also removed.

Signed-off-by: Alexandra Stroiu <alexandra.stroiu@stud.acs.upb.ro>