03-rwslotmachine: Add Dockerfile and Makefile for 03-rwslotmachine dockerization
Add a multi-stage Dockerfile that builds the 32bit binary for
the rwslotmachine3.c file and runs the binary on port 31346.
Local port for build container is 32000 and the
runtime container uses port 32001.

Add a Makefile that handles the Dockerfile's build, run and stop.

It works on my computer. Disk usage seems to be too high.
May need improvements.

Resolves: #50

Signed-off-by: Cosmin Tanasa <tanasa.cosminandrei@gmail.com>